Great_Expectations/get_daily_data/requirements.txt
polygon-api-client
pandas
psycopg2
sqlalchemy
great_expectations==0.18.8